requirements.txt
selenium
selenium.webdriver.common.keys
requests
keyboard
pillow
